Enable calls to `find_dependency` in the exported config only if we are building a static library. (#4526)
#4505 replaced bundling of our dependencies' static libraries with calls
to CMake's `find_dependency` command. We definitely need these calls
when we have installed a static library, and we definitely don't need
them in the release artifacts where everything is contained in a shared
library.

Out of abundance of caution for potential cases where we might need
these options in future scenarios, I controlled emitting these calls
with `TILEDB_INSTALL_FIND_DEPENDENCIES`, an on-by-default option that
was always enabled when building a static library. This caused
disruptions in downstream projects such as SOMA because of this change
in defaults.

My initial thought was to make this option off-by-default if we are not
building a static library, but as it turns out we never need it enabled
in such cases, because a shared library does not require linking to its
dependencies when we link to it, and these dependencies are
implementation details and not exposed through our public C API.
Contrast this with cases like `azure-storage-blobs-cpp` depending on
`azure-storage-common-cpp` at install-time, because the former's public
headers depend on the latter's.

So this PR removes the `TILEDB_INSTALL_FIND_DEPENDENCIES` option and
replaces it with `TILEDB_STATIC`.

---
TYPE: NO_HISTORY