Refactor the group code. (#4378)
[SC-37180](https://app.shortcut.com/tiledb-inc/story/37180/finish-group-refactor-pr)

This PR refactors the group code to reduce state mutations. More
specifically, before writing a group to storage we were calling the
`GroupDetails::apply_pending_changes` method which updated the group
details' canonical store of members. This PR removes
`apply_pending_changes` and replaces it with a `members_to_serialize`
method, that computes the members to write to storage without modifying
the group details object. This makes the code easier to understand.

---
TYPE: IMPROVEMENT
DESC: Refactor the group code.