Update `install-azurite.sh` to not install Node.js and get the latest version of Azurite. (#4535)
Hopefully it fixes CI failures when installing Node.js with apt.

All GitHub Actions images contain a recent version of Node.js either
way, and we can expect developers to install it themselves if they need
to run Azurite. We print a descriptive error message if `npm` is not
found, directing to install Node.js. We also remove the pin to specific
Azurite versions, matching the behavior on Windows.

The Windows script was not updated since it does not pose a problem at
the moment.

---
TYPE: NO_HISTORY