Fix array schema when time traveling
We were using the latest array schema even while traveling. This changes
the latest array schema selection when in ArrayDirectoryMode::READ to
use the newest array schema that is less than or equal to the current
end timestamp. If no array schemas are newer than the timestamp end
value, the earliest schema is used.