Add sm.consolidation.total_buffer_size setting
This adds a new configuration parameter for directly setting the total
buffer size to use during fragment consolidation. This deprecates the
old sm.consolidation.buffer_size value. We will continue to use the old
value if set by users for backwards compatibility. If not set, the new
setting is used.