Update the AWS SDK to version 1.11.215. (#4551)
[SC-37841](https://app.shortcut.com/tiledb-inc/story/37841/update-aws-sdk-to-at-least-1-11-179)

This PR updates the AWS SDK to its latest version as of now (1.11.215).
The first commit copies the ports from the latest version on vcpkg
(1.11.201), and the second commit updates them to their latest version.
The reason for the second update is that version 1.11.205 upstreamed
`remove-checked-array-iterator.patch` that fixes MinGW build failures,
and I found the opportunity to remove the patch from our code.

The ports on vcpkg will be updated with microsoft/vcpkg#35445.

Verified by successfully locally building on my machine. The SHA-512
hash for the `s2n` sources (which is not supported on Windows) was
computed by manually downloading and hashing it.

---
TYPE: IMPROVEMENT
DESC: Update the AWS SDK to version 1.11.215.