For interim comment only: resource management
This PR is the start of a system for management of resources at a whole-library scale. Management here specifically refers to estimation, budgeting, tracking, and enforcement of resources. The first resource to go under such management is memory. It's the only resource in this PR, but won't be the only one.