Fix MemoryTracker lifetime issues
Now that the memory tracker is referenced by allocators, we must ensure
that it outlives all references used. This can be surprisingly tricky
when we consider the possibility of unordered deallocation of various
containers. For instance, one easily observed situation is when we use
the C API to allocate/deallocate an array object in unit tests while
taking references to the FragmentMetadata. The issue is that our
FragmentMetadata destructor can easily run after the call to
`tiledb_array_free` which causes a segfault with the current setup.