Fix aggregate leak when field doesn't exist. (#4610)
When a field didn't exist, we would throw on the creation of the query
field handle after making a handle for the channel. This handle wouldn't
get broken after we throw. This solves the issue by creating the handle
for the channel after we validated that the field exists.

---
TYPE: BUG
DESC: Fix aggregate leak when field doesn't exist.