Refactor fragment consolidation with `class FragmentConsolidationWorkspace` (#4605)
This PR puts the query buffers used inside the fragment consolidator
into their own class. This class deletes its copy constructor but
provides a move constructor. This avoids copying the buffers in the
`return` statement of `create_buffers`, which not only doubles the
instantaneous memory consumption during that statement, but does a bunch
of spurious copying of empty buffers.

In addition, this change prepares for memory as a managed resource,
providing a single location for the allocators of the containers for
these buffers to be specified.

---
TYPE: NO_HISTORY
DESC:  `class FragmentConsolidationWorkspace`