Pass AWS client configuration to `STSProfileWithWebIdentityCredentialsProvider`. (#4641)
This PR applies a similar fix to #4616 but for
`STSProfileWithWebIdentityCredentialsProvider`. That class was also
adjusted to prevent a memory leak, to use the user-provided `STSClient`
if available, and to use public APIs.

The issue has been validated to be fixed.

---
TYPE: BUG
DESC: Fix HTTP requests for AWS assume role with web identity not
honoring config options.