Improve large dense aggregate reads with tile metadata only. (#4657)
The dense reader was creating a bitmap to compute all aggregate results
at the top level of the read. For large aggregate reads where we don't
need to load any tiles, this could be quite large. It also would turn
out to be completely unnecessary. This fix moves the bitmap to the lower
level of the read, where a smaller bitmap can be created only if
necessary.

---
TYPE: IMPROVEMENT
DESC: Improve large dense aggregate reads with tile metadata only.