Move more FragmentMetadata members to tdb::pmr
These particular member variables were chosen because these are the only
variables tracked by the existing MemoryTracker implementation. Once I
implement the counting and switch the implementations it will break a
decent number of unit tests because the new PMR implementations track
all allocations, not just a calculation of what we logically need. In
other words, we're not currently keeping an accurate count of memory,
and getting more accurate counts breaks the existing tests.