Add compression features in the libmagic port.
On Linux it found somewhere a liblzma header and so enabled support for it, but could not link to liblzma.
Now if the specific compression feature is not enabled, support will be explicitly disabled when configuring.
Because we don't use the `MAGIC_COMPRESS` flag anywhere, we don't have to enable these features.