Require shared_ptr for Tile and WriterTile
This moves the Tile and WriterTile classes to requiring that they be
created in a shared_ptr. The motivation for this is that after we start
tracking data allocated in tiles, we would very much like to avoid
having to ensure that our move operations don't actually copy data. This
gets trickier after we start tracking memory via container allocators
because if the allocators don't match we'll end up accidentally copying
instead of moving.

Rather than risk accidentally copying, I've just moved all instances of
the Tile and WriterTile into a shared_ptr for each type.