Implement MemoryTrackerManager
This adds a new MemoryTrackerManager to the ContextResources. All
MemoryTrackers should be created through this API. Eventually I'll
probably make MemoryTracker constructors private so that we force this
condition. However, that would currently make the test changes fairly
verbose so I'm avoiding it for now to reduce the size of this already
decently large diff.