Implement MemoryTrackerReporter
This adds a new MemoryTrackerReporter. If configured, this reporter
starts a background thread that writes a JSON string to the configured
file once a second. To configure the filename, just pass a config value
for `sm.memory.tracker.reporter.filename` when creating the context.
This configuration key is purposefully undocumented so that we aren't
creating a public API contract as the contents of the JSON blobs will
change without warning.