Fix stats bug in StorageManager
I just stumbled across the fact that Serializer::size() returns
semantically different values depending on whether its in size computing
mode vs serialization mode. In size computing mode, it returns the
number of bytes "written" which is the required buffer size. In actual
serializing stuff mode, it returns how much of the buffer is left to be
written, so generally should return 0 after a complete serialization.

Just to see if anyone was misusing that I moved size() to the size
computing serializer and lo and behold, there was a single case where we
had buggy behavior. Fix is to just record the size calculated by the
size computing serializer.