Don't use non empty domain on empty array.
This was introduced in https://github.com/TileDB-Inc/TileDB/pull/4710. In that PR, we could try to use the non empty domain of an array with no fragments, which is empty. This should fix python and R nightlies.

---
TYPE: NO_HISTORY
DESC: Don't use non empty domain on empty array.