TemporaryLocalDirectory changes for no PRNG
I'm poking at trying to separate out the tests in tiledb_unit that can
be reused to test against REST and all our storage backends
transparently. I just spent about an hour re-implementing the PRNG
implementation in TemporaryLocalDirectory since PRNG isn't a public API.

And then about as soon as I got it working I realized that local
filesystem anything in tests is anathema to the whole idea of the test
suite I'm attempting to create. Alas.

Anyway, I did figure out how to properly seed the mt19937_64 generator
if we want to go that far.