Replace boost::container with cpp17::pmr
This removes the boost dependency to avoid having to make that work in
our builds that are still using external packages. This change uses
std::pmr when it is detected as available and usable. On platforms that
don't have a working std::pmr (notably, macOS < 14), it uses the
implementation from Pablo Halpern that is implemented with purely C++11
features.

The major downfall to this approach is that if/when we start changing
allocator strategies it'll require us to backport/reimplement those
allocator strategies for macOS < 14. However, if we end up writing our
own allocation strategies they'll work on either implementation
seamlessly.