Support reading V1 group details with explicit version in the name. (#4744)
[SC-41275](https://app.shortcut.com/tiledb-inc/story/41275/fix-v1-groups-written-with-latest-version)

When version 2 of the group details format was introduced in #3928 (2.16
and also backported to 2.15), the names of group details V2 files were
of the format `__t1_t2_id_version`, and V1 kept using `__t1_t2_id` for
compatibility. In #4530 (2.19), V1 files started being written with the
new format as well. This confused the group reader, which considered
group details files with names of the form `__t1_t2_id_1` as V2, merely
because they had a version field in their name.

This PR fixes the group reader, making it always consider the version if
it is present in the name of details file. The change has been validated
by writing to a V1 group and then reading from it. The test group was
created locally by Core version 2.14 and was added in
`test/inputs/groups` (its size is tiny so it can safely be stored in the
repository).

---
TYPE: BUG
DESC: Support reading V1 group details with explicit version in the
name.