Update crop_range to clamp to domain range. (#4781)
While working on #4685 we found it was possible for a cropped range to
fall outside of the given domain range. This updates crop_range to call
`std::clamp` instead of min/max to ensure the resulting cropped range is
within the domain.

---
TYPE: BUG
DESC: Update crop_range to clamp to domain range.