Remove QueryCondition requirement for Enumeration values to be valid. (#4707)
This change removes the requirement that values used in comparisons
against enumerated attributes exist as an value of an enumeration. That
is to say, previously if your enumerations were a list of colors, and
you attempted to create a query condition of `value =
"something_not_a_color"` TileDB would raise an exception saying that
`"something_not_a_color"` was not a valid color value.

Removing that requirement also means that we have to update query
condition logic to handle various situations when a condition will never
match or always match. In order to handle that logic, this PR adds to
new QueryCondition "operators" `ALWAYS_TRUE` and `ALWAYS_FALSE`. These
operators only exist internally and can't be referenced or created
outside of the C API boundary. These operations are inserted while a
query is rewritten which only happens internally to a query execution.

When using a query condition with a value that is not a member of a
enumeration there are three distinct logical situations. For every
binary operator except `!=` the result is `ALWAYS_FALSE`. Obviously, a
value of the enumeration can never equal or be ordered against an
unknown value. The `!=` is special given the negation means *every*
enumerated value matches since by definition they aren't equal to an
unknown value. The third case is for the `IN` and `NOT_IN` operations.
These don't have logical changes, we just don't add the unknown value to
the set of values to test against.

---
TYPE: IMPROVEMENT
DESC: Remove QueryCondition requirement for Enumeration values to be
valid.