Fix segfault in ArraySchema::check
Previously we weren't accounting for when enumerations weren't loaded
and a check was performed. This lead to attempts to dereference the
nullptr when the enumeration hadn't been loaded.

Given that enumerations are only unloaded when reading an array, that
means these checks were already run when the schema was written so
ignoring them here is fine since they're just enforcing a maximum size
constraint.