SC 7826: Convert Dimension function pointer members to subclass methods (#4774)
The `Dimension` class has 20+ function pointer members which are set to
dynamically dispatch a template instantiation of a function based on the
data type of the dimension. The `Dimension` constructor sets each of
these function pointers, switching on the type for each of them. But
really the same switch case applies to all of them. It is a cleaner
encapsulation of functionality to nest the implementations within a
subclass, which also adds a small benefit of reducing the object size.

Hence this pull request introduces `DimensionDispatch`, which takes the
function pointers we previously had and makes them pure virtual methods
of a base class instead. `template <typename T> DimensionDispatchBase`
implements most of these virtual methods, and defers any exceptions to
the pattern to `DimensionFixedSize` and `DimensionVarSized`. And the
`Dimension` constructor chooses an implementation with a single function
call `set_dimension_dispatch` instead of the previous logic to set each
function pointer individually.

Closes sc-7826

---
TYPE: IMPROVEMENT
DESC: Convert dimension function pointer members to subclass methods