Support VFS ls_recursive API for posix filesystem. (#4778)
@shaunrd0 previously added `ls_recursive` support for the S3 file
system. This pull request implements it for Posix, in a
hopefully-portable way (`std::filesystem::recursive_directory_iterator`)
which can be recycled for Windows.

Some notes about input/output:
- while `LsObjects` has a `std::string` component, the string is a URI.
- for posix, we include empty directories in the result set.

The recursion is implemented so that if a directory does not pass the
directory filter, then we do not descend into that directory. I have
added a unit test to check this behavior. Something like this could be
useful for S3 as well if the "list objects" request is done in
hierarchical mode.

---
TYPE: FEATURE
DESC: Support VFS `ls_recursive` API for posix filesystem.

---------

Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>