Add a test for concurrent array metadata writes. (#4806)
Add a test for concurrent array metadata writes. With the fragment name
generation, this test is tagged as `[!mayfail]`, because there's no
means of sorting fragments which are written within the same
millisecond. Once sub millisecond temporal disambiguation is
implemented, this test should serve as a regression case.

---
TYPE: NO_HISTORY
DESC: Add a test for concurrent array metadata writes.

---------

Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>