Do not load group metadata when getting it from REST.
This was introduced in https://github.com/TileDB-Inc/TileDB/pull/4760 where the client tries to load the group metadata when serializing the group metadata request to REST, which causes an infinite loop. The fix reverts to the correct behavior, where we don't try to load the metadata when serializing it from the client.

---
TYPE: BUG
DESC: Do not load group metadata when getting it from REST.