Add new and delete to DataBlock so it can be used with std::allocator (#4075)
This PR changes the notion of `DataBlock` slightly. We now consider the
`DataBlock` to be a single thing that is allocated. According, we no
longer consider `PoolAllocator` as an actual allocator. Rather, we
overload `new` and `delete` in `DataBlockImpl` so that `DataBlock`s can
be used directly with `std::allocator`.

---
TYPE: NO_HISTORY
DESC: Add new and delete to DataBlock so it can be used with
std::allocator

---------

Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>