Fix timestamp usage with random label. (#4830)
Random label uses the current time to know if a new label needs to use
an incrementing counter so that labels generated in the same millisecond
sort properly when reading back. Since the code at the higher level used
a timestamp and the random label used another, it was possible to end up
with timestamped names having the same millisecond, but not sorting
properly. The fix is to have the random label code return the timestamp
it used to the caller so that the caller can use it to generate
timestamped names.

---
TYPE: NO_HISTORY
DESC: Fix timestamp usage with random label.