Improve diagnostics when an Azure endpoint is not configured. (#4845)
[SC-43870](https://app.shortcut.com/tiledb-inc/story/43870/remove-azure-warning-when-not-using-azure-credentials)

This PR changes the Azure VFS to fail with an explicit error if an
operation on the Azure VFS was attempted, but an endpoint to Azure Blob
Storage could not be determined based on the
`vfs.azure.storage_account_name` and `vfs.azure.blob_endpoint`
configuration options. The previous behavior was to emit a warning,
which happened at the initialization of the VFS class if Azure is
enabled, regardless of whether Azure is being actually used. That
warning is no longer emitted.

---
TYPE: IMPROVEMENT
DESC: Improve diagnostics when an Azure endpoint is not configured.