Always use vcpkg. (#4570)
[SC-36911](https://app.shortcut.com/tiledb-inc/story/36911/remove-enable-vcpkg-options-from-cmake-command-line-and-bootstrap-scripts)

This PR changes the severity of the message when we configure with
`-DTILEDB_VCPKG=OFF` from a `DEPRECATION` to a `FATAL_ERROR`. This means
that using the legacy `ExternalProject`-based mechanism for acquiring
dependencies is no longer supported. The `ExternalProject` modules have
become dead code and will be removed in a subsequent PR.

---
TYPE: BUILD
DESC: Vcpkg is always enabled; turning the `TILEDB_VCPKG` option off is
no longer supported and fails.