Fix build errors with CMake 3.29.1. (#4859)
CMake 3.29.1 had a change that broke vcpkg's `vcpkg_cmake_config_fixup`
command. This PR updates the port containing the command to adapt to the
new behavior. The change will be upstreamed with microsoft/vcpkg#38017.

Validated locally. For ease of review, the first commit copies the
existing port as-is, and the second commit updates it.

Fixes #4857
Fixes TileDB-Inc/TileDB-CSharp#405
Fixes the root cause of TileDB-Inc/conda-forge-nightly-controller#83

---
TYPE: NO_HISTORY