Stabilize REST CI (#4876)
Moves generators in the beginning of tests. Flow of execution and clean
up can get unclear/messy when GENERATORS and SECTIONs are interleaved.

Testing: I re-ran the REST-CI 3 times and it passed all of them. Not
sure the problem is fixed, but for sure it didn't make it worse.

sc-42190

---
TYPE: NO_HISTORY
DESC: Stabilize REST CI