Add end-to-end cloud test workflow with Microsoft Entra ID. (#4841)
This PR adds the foundations for testing connectivity of TileDB to the
real clouds by adding a new workflow named `test-cloud-e2e`. Currently
only Azure with Microsoft Entra ID authentication is provided.

Authentication happens [with OpenID
Connect](https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-azure).
A new environment was created that has read-only access to the `core`
container of the `tiledbci` storage account, and we test that the file
`test.txt` in that container exists. The test is configured with
environment variables, making it reusable for subsequent tests with
other cloud providers.

---
TYPE: NO_HISTORY

---------

Co-authored-by: Theodore Tsirpanis <theodore.tsirpanis@tiledb.com>