Move the spdlog header files to `logger.cc`. (#4881)
[SC-44926](https://app.shortcut.com/tiledb-inc/story/44926/move-spdlog-header-files-to-logger-cc)

Spdlog includes `Windows.h` on Windows which includes lots of code,
including some problematic defines like `DELETE`. This PR refactors the
`Logger` class to no longer need including spdlog. We accomplish this by
forward-declaring the `spdlog::logger` class and updating the formatted
logging template methods to format the string with `fmt` themselves, and
pass the string to the overload that logs strings. Formatting the
strings is avoided if the respective log level is enabled. I added a
method to check that.

---
TYPE: NO_HISTORY