Add C++ API for tiledb_array_consolidate_fragments. (#4884)
Adding a missing C++ API. 

To pass the test I created I needed to use the last part of the full
fragment URIs which I got from the FragmentInfo. I believe I should have
been able to use the URIs as is, without further modifications. This is
happening because of:


https://github.com/TileDB-Inc/TileDB/blob/6dff6fae0b09c16177fe85ec9b62c128bd721ad7/tiledb/sm/consolidator/fragment_consolidator.cc#L377

I decided to keep this PR just a C++ addition. I can follow-up with an
investigation if the behavior I describe above is not the expected.

[sc-44902]

---
TYPE:  CPP_API 
DESC: Add C++ API for tiledb_array_consolidate_fragments.