C API support - to_string_view
Add function `tiledb::api::to_string_view()`. This function validates a candidate C-style string and converts it to a `std::string_view`. It's specific to the C API, with a template argument that allows an error message to identify an erroneous argument.

Converted C API functions `tiledb_query_get_est_result_size*` to use this function. Changed relevant functions in `class Query` and `class Subarray` to use `string_view` and not a raw `char *`.