Expand `TILEDB_REMOVE_DEPRECATIONS` to exclude all deprecated C and C++ APIs. (#4887)
[SC-23937](https://app.shortcut.com/tiledb-inc/story/23937/build-artifact-disabling-deprecated-functions)

> [!NOTE]
> Depends on #4879.

As part of preparing to remove the deprecated APIs, we need a mechanism
to exclude them from being compiled. There exists already the
`TILEDB_REMOVE_DEPRECATIONS` option, but it covers only a subset of the
deprecated APIs in the C++ `Array` class.

This PR expands `TILEDB_REMOVE_DEPRECATIONS` to cover every deprecated C
and C++ API. When this option is set, a define with the same name is
defined and the deprecated APIs are not compiled into the library. This
flag also gets publicly exported from the `TileDB::tiledb` target.

---
TYPE: BUILD
DESC: Update the `TILEDB_REMOVE_DEPRECATIONS` option to exclude all
deprecated C and C++ APIs from the library binary and the headers.