Implement permutation view for external sort.
This pull request implements the `permutation_view` class.  A `permutation_view` provides a view over an underlying range that is permuted by indirecting with a given index range.  It is required that the permutation range is the same size as the data range and that it  contain the values 0, 1, ..., N-1, in arbitrary order, where N is the size of the data range.

Example:
```c++
  std::vector<int> v{1, 2, 3, 4, 5};
  std::vector<size_t> p{4, 3, 2, 1, 0};
  permutation_view view(v, p);

  CHECK(std::ranges::equal(view, std::vector<int>{5, 4, 3, 2, 1}));
```

The implementation uses `iterator_facade` to realize a C++20 standards conformant iterator and so that `permutation_view` satisfies the appropriate C++20 ranges concepts.

The `permutation_view` class has been tested with `var_length_view`.  Various range and iterator concepts are also tested.

(This PR also cleans up the documentation and the structure of the `var_length_view` class.)

---
TYPE: IMPROVEMENT
DESC: Implement permutation view for external sort.