Implement proxy sort for external sort. (#4922)
This PR provides four overloads of a "proxy sort" algorithm and four
overloads of a stable "proxy sort". Instead of directly sorting a range
`x` such that `x[I] <= x[I+1]`, a proxy sort sorts a permutation range
`p` so that `x[p[I]] <= x[p[I+1`]]`.

 Four variations of proxy_sort and stable_proxy_sort are provided
* `proxy_sort_no_init`: Sorts the permutation, given an initial
permutation. By default, < is used to compare, but an overload is
provided that takes a comparison function.

* `proxy_sort`: Sorts the permutation, initializing it to the identity.
By default, < is used to compare, but an overload is provided that takes
a comparison function.

* `stable_proxy_sort_no_init`: Performs stable sort of the permutation,
given an initial permutation. By default, < is used to compare, but an
overload is provided that takes a comparison function.

* `stable_proxy_sort`: Performs stable sort of the permutation,
initializing it to the identity. By default, < is used to compare, but
an overload is provided that takes a comparison function.

Example:
```c++
 std::vector<int> x{3, 1, 4, 1, 5, 9, 2, 6, 5, 3};
 std::vector<int> perm(x.size());
 std::vector<int> expected{1, 1, 2, 3, 3, 4, 5, 5, 6, 9};
 proxy_sort(x, perm);
 for (size_t i = 0; i < x.size(); ++i) {
   CHECK(x[perm[i]] == expected[i]);
 }
```

---
TYPE: IMPROVEMENT
DESC: Implement proxy sort for external sort.