C API support - to_string_view (#4915)
Add function `tiledb::api::to_string_view()`. This function validates a
candidate C-style string and converts it to a `std::string_view`. It's
specific to the C API, with a template argument that allows an error
message to identify an erroneous argument.

Converted C API functions `tiledb_query_get_est_result_size*` to use
this function. Changed relevant functions in `class Query` and `class
Subarray` to use `string_view` and not a raw `char *`.

This is a followup to #4891, where this work was out of scope. Had the
conversion function already existed before that PR, it would have used
there.

---
TYPE: NO_HISTORY
DESC: C API support - to_string_view

---------

Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>