Additional constructors for external sort classes. (#4928)
This PR is stacked on #4925. It adds additional constructors for the two
view classes for var length data.

The previous implementations had constructors that just took input
ranges, and constructed a var length view over the data specified by
those ranges. However, in general, we may have a range that is not
completely filled with data, and so may need to specify the data to be
viewed in some other way. The complete set of constructors take
* ranges for data and offsets (arrow format) - the size of the resulting
view is one less than the size of the offset range
* ranges plus sizes for data and offsets (arrow format) - the size of
the resulting view is one less than the size given for the offset range
* iterator pairs for data and offsets (arrow format) - the size of the
resulting view is one less than the difference between end and begin of
the offsets pair
* iterator pairs for data and offsets, with sizes (arrow format) - the
size of the resulting view is one less than the size given for the
offset range
* ranges for data and offsets (tilledb format) - the size of the
resulting view is equal to the size of the offset range (takes an extra
argument for last offset value)
* ranges plus sizes for data and offsets (tiledb format) - the size of
the resulting view is equal to the size given for the offset range
(takes an extra argument for last offset value)
* iterator pairs for data and offsets (tiledb format) - the size of the
resulting view is equal to the difference between end and begin of the
offsets pair (takes an extra argument for last offset value)
* iterator pairs for data and offsets, with sizes (tiledb format) - the
size of the resulting view is equal than the size given for the offset
range (takes an extra argument for last offset value)

[sc-44576]

---
TYPE: NO_HISTORY
DESC: Additional constructors for external sort classes.