Implement zip_view for external sort. (#4930)
This PR implements a `zip_view` for zipping together a set of ranges. It
is intended to implement the `std::ranges::zip_view` as defined for
C++23. From
 https://en.cppreference.com/w/cpp/ranges/zip_view:
1. A zip_view is a range adaptor that takes one or more views, and
produces a view whose ith element is a tuple-like value consisting of
the ith elements of all views. The size of produced view is the minimum
of sizes of all adapted views.
2. `zip()` is a customization point object that constructs a `zip_view`

Currently, the `zip_view` only supports zipping together ranges that are
`random_access_range`s. In addition, the `size()` and `end()` functions
are only provided if all of the ranges are `sized_range`s

The iterator from a `zip_view` is essentially a tuple of pointers to the
beginning of each of the zipped ranges, plus an index that keeps track
of the iterator's location in the zipped ranges. The size of a
`zip_view` is the size of the smallest range. The end iterator of a zip
view is the begin iterator plus the size of the zip_view.

Unit tests have similar coverage to the tests for the var length views.
Tests also include zipping a var length view and iterating through with
`std::for_each` and `for`.

[sc-43639]

---
TYPE: IMPROVEMENT
DESC: Implement zip_view for external sort.