Implement proxy sort permutation for external sort.
This PR implements proxy sorting for `permutation_view`.  The goal is to be able to efficiently sort a `zip_view` containing a number of ranges.

With the PR, we can sort a zipped `permutation_view` as follows:
```c++
  auto z = zip(u, v, w, x);
  auto p = permutation_view(z);
  sort(p);
```
resulting in `z` being sorted along `u` (then `v`, etc).

Because the iterator for `permutation_view` only returns a value, with no information about the index, we can't sort a `permutation_view` directly with, e.g., `std::sort`.  To enable sorting, we instead provide sorting member functions to the `permutation_view` and have free functions that are overloaded on `permutation_view`  invoke the member functions.

The functions added include
* proxy sort member functions in `permutation_view` with `_no_init`, `stable_`, and comparison function variants.
* proxy sort free functions overloaded on `permutation_view` with variants as above
* Overloads of `std::sort` and `std::stable_sort`.  These invoke the `no_init` member function, the assumption being that we just want to stack permutations on top of each other.

Since the `proxy_sort` functions mutate the indexes in the `permutation_view`, the interfaces to the existing `proxy_sort` function were updated to take a forwarding reference.

A few notes:
* Sorting a `permutation_view` will mutate the underlying permutation index range.  Right now, the permutation view is just a view over two containers (the one to be permuted and the permutation itself).  Sorting the `permutation_view` will change the ordering of the container passed in to the constructor.
* When sorting a `permutation_view` with a tuple value type, the sorting will be done with the built-in tuple comparator.  We may want to add variants that let us specify which members to sort on (using indexes passed in as variadic template parameters). However, using a specialized comparator is probably good enough.

---
TYPE: IMPROVEMENT
DESC: Implement proxy sort permutation for external sort.